fix(ghanasps): stop the plot-identity script breaking pytest collection
My fault.  The script I added in the previous commit was named
`plot_identity_permutation_test.py`, which matches pytest's DEFAULT
`python_files` pattern `*_test.py`.  This repo has no
`[tool.pytest.ini_options]` block at all -- so no `testpaths`, and
`python_files` is at its default -- and pytest therefore collected it from
the repo root.  It then did data loading at MODULE level, which needs S3
credentials CI does not have, and an error during COLLECTION aborts the
entire session: zero tests ran, not one.  Same shape as GH #680.

Two independent fixes, either of which would have been sufficient:

1. Renamed to `plot_identity_permutation.py` -- matches neither `test_*.py`
   nor `*_test.py`, so pytest never imports it.
2. Moved all I/O behind `if __name__ == "__main__":`, so importing the
   module can never touch the network whatever it is called.

Verified:

- Import with credentials explicitly blanked (`AWS_ACCESS_KEY_ID=`,
  `AWS_SECRET_ACCESS_KEY=`, `AWS_SHARED_CREDENTIALS_FILE=/nonexistent`)
  succeeds and does no I/O.
- Collected in an isolated directory with no conftest: the new content
  collects in 0.39s, the old content under a `*_test.py` name takes
  10.13s -- ten seconds of data loading during collection, which is
  precisely the module-level I/O that raises NoCredentialsError on CI.

Honest caveat: I could not reproduce the CI credential failure itself
locally, because this machine has the source data cached, so the old
version merely ran slowly instead of raising.  The 0.39s / 10.13s split is
the evidence that the module-level I/O is gone; the CI traceback for the
failure mode came from the reviewer.

Also verified there is no other `test_*.py` / `*_test.py` anywhere under
`slurm_logs/`, so this was the only file of mine at risk.

Not touched: the systemic half (adding `testpaths` so the repo root is not
scanned) is being handled separately to avoid a collision.  Note that
`SkunkWorks/` still carries four collectable `test_*.py` files that survive
only because their module-level code happens not to raise.
